Documentación Se edita el README para tener una documentación sobre lo realizado. Resta realizar pruebas pasando valores Strings

+++ src/main/java/com/books/integrate/spring/react/repository/TutorialRepository.java
@@ -11,5 +11,5 @@ public interface TutorialRepository extends JpaRepository<Tutorial, Long> {
 
 	List<Tutorial> findByTitle(String title);
 
-	List<Tutorial> findByPrice(Long price);
+	List<Tutorial> findByPrice(Double price);
 }
